// src/app/common/model/FormTank-class.ts
export class TankForm {
  aqueductName: string;
  registerNo: string;
  inCharge: string;
  tankName: string;
  volumeAmount: string;
  volumeUnit: string;
  creationDate?: string;
  type: string;
  material: string;
  cleaning: string;
  latitude: number;
  longitude: number;
  direction: string;
}